requirements.txt
numpy>=1.24.0
torch>=2.2.0
PyYAML>=6.0
tqdm>=4.66.0
